pyproject.toml
[project]
name = "agentic-cold-emailer-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "faiss-cpu>=1.13.2",
    "google-generativeai>=0.8.6",
    "huggingface-hub>=1.11.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.14",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.2",
    "numpy>=2.4.4",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pinecone-client>=6.0.0",
    "pydantic>=2.13.2",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
]
